Encapsulate benchmark running
Changes to the internal workings of what benchmark runner we
are using and how it works are now transparent to the running
application
